feat(snddrv): add snd_set_p_tables() API
